Renamed command
search -> urban
